third_party/blink/web_tests/platform/fuchsia/http/tests/security/no-javascript-refresh-expected.txt

CONSOLE ERROR: Refused to refresh http://127.0.0.1:8000/security/no-javascript-refresh.php to a javascript: URL
This it the content that appears in place of a refresh.
